Decay the net position instead of tallying forever
bought_amount and sold_amount only ever went up, so the price was a function of
every trade that had ever happened. Nothing in the system pointed back at
base_price. An item that got dumped in month one was still on the floor in
month twelve and always would be, which is the permanently-min-price group.

Net position is now its own value that halves on a per item half-life, settled
lazily on read. No scheduler, nothing to catch up after a restart, one pow()
per lookup. Default 48h: short enough that a crash heals within a session or
two, long enough that impact still means something.

Worth noting decay cannot be farmed in either direction. Sell, wait, buy back
and the price has recovered so it costs more; buy, wait, sell and it has sagged
so it pays less. Recovery always moves against whoever is round tripping.
Both directions are covered in the tests.

The tallies stay as lifetime analytics for item_history. They are BIGINT now —
a cobble farm gets to 2.1bn eventually. Totals are long, per-trade deltas stay
int.

items.sql is CREATE TABLE IF NOT EXISTS so it never reaches an existing
install, and MySQL has no ADD COLUMN IF NOT EXISTS, so migrate() checks the
catalog and seeds net_position from bought - sold on first run.

/dmarket set moves net too. Setting the tallies alone would rewrite the
analytics and leave the price where it was.

+++ src/main/java/me/wuzzyxy/dynamicmarket/database/Database.java
@@ -8,8 +8,8 @@ public interface Database {
     void die();
 
     MarketItem addItem(String item, double basePrice, double minPrice, double percentage);
-    MarketItem addItem(String item, double basePrice, double minPrice, int boughtAmount, int soldAmount, double percentage);
-    MarketItem setItem(String item, double basePrice, double minPrice, int boughtAmount, int soldAmount, double percentage);
+    MarketItem addItem(String item, double basePrice, double minPrice, long boughtAmount, long soldAmount, double percentage);
+    MarketItem setItem(String item, double basePrice, double minPrice, long boughtAmount, long soldAmount, double percentage);
     MarketItem setItemStatics(String item, double basePrice, double minPrice, double percentage);
     boolean removeItem(String item);
     MarketItem getItem(String item);
@@ -24,9 +24,9 @@ public interface Database {
     MarketItem getSoldAmount(MarketItem item);
     MarketItem addBoughtAmount(MarketItem item, int amount);
     MarketItem addSoldAmount(MarketItem item, int amount);
-    MarketItem setBoughtAmount(MarketItem item, int amount);
-    MarketItem setSoldAmount(MarketItem item, int amount);
-    MarketItem setAmounts(MarketItem item, int boughtAmount, int soldAmount);
+    MarketItem setBoughtAmount(MarketItem item, long amount);
+    MarketItem setSoldAmount(MarketItem item, long amount);
+    MarketItem setAmounts(MarketItem item, long boughtAmount, long soldAmount);
 
     @Deprecated
     boolean createHistoryPoint(MarketItem item);
